requirements.txt
  
apscheduler==3.6.3
bs4==0.0.1
certifi==2019.6.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==5.0.3
Django==2.1.11
django-cors-headers==3.1.0
django-enumchoicefield==1.1.0
django-filter==2.2.0
django-mssql==1.8
django-pyodbc-azure==2.1.0.0
djangorestframework==3.10.2
django-rest-swagger==2.2.0
django-separatedvaluesfield==0.6.0
drf-oidc-auth==0.9
fpdf==1.7.2
holidays==0.9.12
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
numpy==1.17.4
openapi-codec==1.3.2
pandas==0.25.1
pdfkit==0.6.1
pyodbc==4.0.27
pytz==2019.2
requests~=2.24.0
simplejson==3.16.0
sqlparse==0.3.0
translate==3.5.0
uritemplate==3.0.0
urllib3==1.25.3
workalendar==7.2.0
openpyxl==3.0.3

XlsxWriter~=1.3.2
